requirements.txt
fastapi>=0.100
uvicorn>=0.20
fin123-core @ git+https://github.com/reckoning-machines/fin123-core.git@ed7e210a3a82f4a6545207aead325d05c5cd04f4
